split GetTransactionsForOrdering to its own file
