+++ services/transactionpool/validate_transaction_test.go
@@ -61,4 +61,3 @@ func TestValidateTransaction_InvalidTransactions(t *testing.T) {
 		})
 	}
 }
-
